Update Dashboard page
