+++ packages/frontend/app/(application)/page.tsx
@@ -1,18 +1,13 @@
 import { Metadata } from 'next';
 
-import { Layout } from '$layouts/Layout';
-import { Dashboard } from '$views/dashboard/dashboard';
+import { DashboardContainer } from '$container/DashboardContainer';
 
 export const metadata: Metadata = {
   title: 'Dashboard',
 };
 
 const HomePage = () => {
-  return (
-    <Layout title="Dashboard">
-      <Dashboard />
-    </Layout>
-  );
+  return <DashboardContainer />;
 };
 
 export default HomePage;
